requirements.txt
aniso8601==9.0.1
argon2-cffi==20.1.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coverage==5.5
Flask==2.0.1
Flask-RESTful==0.3.9
freezegun==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
